OCR page: missing case fixed + file splitted
