+++ packages/smooth_app/lib/pages/product/edit_ocr/edit_ocr_page.dart
@@ -15,12 +15,14 @@ import 'package:smooth_app/helpers/analytics_helper.dart';
 import 'package:smooth_app/helpers/product_cards_helper.dart';
 import 'package:smooth_app/pages/image_crop_page.dart';
 import 'package:smooth_app/pages/product/common/product_refresher.dart';
+import 'package:smooth_app/pages/product/edit_ocr/ocr_helper.dart';
 import 'package:smooth_app/pages/product/explanation_widget.dart';
 import 'package:smooth_app/pages/product/multilingual_helper.dart';
-import 'package:smooth_app/pages/product/ocr_helper.dart';
 import 'package:smooth_app/pages/product/product_image_button.dart';
 import 'package:smooth_app/themes/smooth_theme_colors.dart';
 import 'package:smooth_app/widgets/smooth_scaffold.dart';
+
+part 'edit_ocr_main_action.dart';
 
 /// Editing with OCR a product field and the corresponding image.
 ///
@@ -76,8 +78,6 @@ class _EditOcrPageState extends State<EditOcrPage> with UpToDateMixin {
         return;
       }
 
-      setState(() => _extractingData = false);
-
       if (extractedText == null || extractedText.isEmpty) {
         await LoadingDialog.error(
           context: context,
@@ -89,7 +89,10 @@ class _EditOcrPageState extends State<EditOcrPage> with UpToDateMixin {
       if (_controller.text != extractedText) {
         setState(() => _controller.text = extractedText);
       }
-    } catch (_) {}
+    } catch (_) {
+    } finally {
+      setState(() => _extractingData = false);
+    }
   }
 
   /// Updates the product field on the server.
@@ -376,223 +379,15 @@ class _EditOcrPageState extends State<EditOcrPage> with UpToDateMixin {
     return result;
   }
 
-  OcrState _extractState(TransientFile transientFile) {
+  _OcrState _extractState(TransientFile transientFile) {
     if (_extractingData) {
-      return OcrState.EXTRACTING_DATA;
+      return _OcrState.EXTRACTING_DATA;
     } else if (transientFile.isServerImage()) {
-      return OcrState.IMAGE_LOADED;
+      return _OcrState.IMAGE_LOADED;
     } else if (transientFile.isImageAvailable()) {
-      return OcrState.IMAGE_LOADING;
+      return _OcrState.IMAGE_LOADING;
     } else {
-      return OcrState.OTHER;
+      return _OcrState.OTHER;
     }
   }
-}
-
-class _EditOcrMainAction extends StatelessWidget {
-  const _EditOcrMainAction({
-    required this.onPressed,
-    required this.helper,
-    required this.state,
-  });
-
-  final VoidCallback onPressed;
-  final OcrHelper helper;
-  final OcrState state;
-
-  @override
-  Widget build(BuildContext context) {
-    final AppLocalizations appLocalizations = AppLocalizations.of(context);
-
-    final Widget? child = switch (state) {
-      OcrState.IMAGE_LOADING => _EditOcrActionLoadingContent(
-          helper: helper,
-          appLocalizations: appLocalizations,
-        ),
-      OcrState.IMAGE_LOADED => _ExtractMainActionContentLoaded(
-          helper: helper,
-          appLocalizations: appLocalizations,
-          onPressed: onPressed,
-        ),
-      OcrState.EXTRACTING_DATA => _EditOcrActionExtractingContent(
-          helper: helper,
-          appLocalizations: appLocalizations,
-        ),
-      OcrState.OTHER => null,
-    };
-
-    if (child == null) {
-      return EMPTY_WIDGET;
-    }
-
-    final SmoothColorsThemeExtension theme =
-        Theme.of(context).extension<SmoothColorsThemeExtension>()!;
-
-    return SizedBox(
-      height: 45.0 * (_computeFontScaleFactor(context)),
-      width: double.infinity,
-      child: DecoratedBox(
-        decoration: BoxDecoration(
-          borderRadius: ANGULAR_BORDER_RADIUS,
-          color: theme.primarySemiDark,
-          border: Border.all(
-            color: theme.primaryBlack,
-            width: 2.0,
-          ),
-        ),
-        child: Material(
-          type: MaterialType.transparency,
-          child: ProgressIndicatorTheme(
-            data: const ProgressIndicatorThemeData(
-              color: Colors.white,
-            ),
-            child: IconTheme(
-              data: const IconThemeData(color: Colors.white),
-              child: DefaultTextStyle(
-                style: const TextStyle(
-                  fontWeight: FontWeight.bold,
-                  fontSize: 15.5,
-                  color: Colors.white,
-                ),
-                child: child,
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  double _computeFontScaleFactor(BuildContext context) {
-    final double fontSize = DefaultTextStyle.of(context).style.fontSize ?? 15.0;
-    final double scaledFontSize =
-        MediaQuery.textScalerOf(context).scale(fontSize);
-
-    return scaledFontSize / fontSize;
-  }
-}
-
-class _EditOcrActionExtractingContent extends StatelessWidget {
-  const _EditOcrActionExtractingContent({
-    required this.helper,
-    required this.appLocalizations,
-  });
-
-  final OcrHelper helper;
-  final AppLocalizations appLocalizations;
-
-  @override
-  Widget build(BuildContext context) {
-    return Semantics(
-      label: helper.getActionLoadingPhoto(appLocalizations),
-      excludeSemantics: true,
-      child: Shimmer(
-        gradient: const LinearGradient(
-          begin: Alignment.centerLeft,
-          end: Alignment.centerRight,
-          colors: <Color>[
-            Colors.black,
-            Colors.white,
-            Colors.black,
-          ],
-        ),
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: MEDIUM_SPACE),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: <Widget>[
-              const CircularProgressIndicator.adaptive(),
-              Expanded(
-                child: Text(
-                  helper.getActionExtractingData(appLocalizations),
-                  textAlign: TextAlign.center,
-                ),
-              ),
-              const CircularProgressIndicator.adaptive(),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _ExtractMainActionContentLoaded extends StatelessWidget {
-  const _ExtractMainActionContentLoaded({
-    required this.helper,
-    required this.appLocalizations,
-    required this.onPressed,
-  });
-
-  final OcrHelper helper;
-  final AppLocalizations appLocalizations;
-  final VoidCallback onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    return Semantics(
-      excludeSemantics: true,
-      value: helper.getActionExtractText(appLocalizations),
-      button: true,
-      child: InkWell(
-        onTap: onPressed,
-        borderRadius: ANGULAR_BORDER_RADIUS,
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: MEDIUM_SPACE),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: <Widget>[
-              Expanded(
-                child: Text(
-                  helper.getActionExtractText(appLocalizations),
-                ),
-              ),
-              const Icon(
-                Icons.download,
-                semanticLabel: '',
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _EditOcrActionLoadingContent extends StatelessWidget {
-  const _EditOcrActionLoadingContent({
-    required this.helper,
-    required this.appLocalizations,
-  });
-
-  final OcrHelper helper;
-  final AppLocalizations appLocalizations;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: MEDIUM_SPACE),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: <Widget>[
-          Expanded(
-            child: Text(
-              helper.getActionLoadingPhoto(appLocalizations),
-            ),
-          ),
-          const CircularProgressIndicator.adaptive(),
-        ],
-      ),
-    );
-  }
-}
-
-enum OcrState {
-  IMAGE_LOADING,
-  IMAGE_LOADED,
-  EXTRACTING_DATA,
-  OTHER,
 }
