+++ packages/smooth_app/lib/pages/product/edit_ocr/ocr_helper.dart
